Fix how we do intents, make navigator static and add unique id to each intent
