+++ sdk/src/main/java/com/stytch/sdk/Util.kt
@@ -1,25 +1,13 @@
 package com.stytch.sdk
 
+import android.app.Activity
+import android.content.Context
 import android.content.Intent
-import android.util.Base64
 import android.util.Patterns
 import java.io.Serializable
 
 internal fun String.isValidEmailAddress(): Boolean {
     return Patterns.EMAIL_ADDRESS.matcher(this).matches()
-}
-
-internal fun generateAuthorizationHeader(projectId: String, secret: String): String {
-    return "Basic " + Base64.encodeToString(
-        "$projectId:$secret".toByteArray(),
-        Base64.NO_WRAP,
-    )
-}
-
-internal fun assertInitialized() {
-    if (!Stytch.isInitialized) {
-        error("Stytch Error: Stytch not initialized. You must call 'Stytch.configure(...)' before using any functionality of the Stytch SDK.")
-    }
 }
 
 internal inline fun <reified T : Serializable> intentWithExtra(extra: T): Intent {
@@ -33,21 +21,27 @@ public sealed class StytchUIResult {
     public object CancelledByUser : StytchUIResult()
 }
 
-internal fun Intent?.toStytchUIResultInternal(internalCall: Boolean = true): StytchUIResult {
+public fun Intent?.toStytchUIResult(): StytchUIResult {
+    disposeAllScreens()
     return when (this) {
         null -> StytchUIResult.CancelledByUser
         else -> {
             val user = getSerializableExtra<StytchResponseTypes.AuthenticateUserResponse>()
-                ?: error(
-                    if (internalCall) "Stytch Error: Internal Error. Could not deserialize user."
-                    else "Stytch Error: Called Intent?.toStytchUIResult on an invalid Intent. Are you checking for the correct request code in onActivityResult()?"
-                )
+                ?: error("Stytch Error: Called Intent?.toStytchUIResult on an invalid Intent. Are you checking for the correct request code in onActivityResult()?")
             StytchUIResult.Success(user)
         }
     }
 }
 
-public fun Intent?.toStytchUIResult(): StytchUIResult = toStytchUIResultInternal(false)
+internal fun <T : Activity> stytchIntent(context: Context, activity: Class<T>): Intent {
+    return Intent(context, activity).apply {
+        putExtra(Constants.ACTIVITY_ID_EXTRA_NAME, StytchActivity.getNextActivityId())
+    }
+}
+
+internal fun disposeAllScreens() {
+    StytchActivity.navigator = null
+}
 
 internal inline fun <reified T : Serializable> Intent.withSerializableExtra(extra: T): Intent = apply {
     putExtra(T::class.qualifiedName, extra)
@@ -60,4 +54,8 @@ internal inline fun <reified T : Serializable> Intent?.getSerializableExtra(): T
 public object StytchErrorTypes {
     public const val EMAIL_NOT_FOUND: String = "email_not_found"
     public const val BILLING_NOT_VERIFIED_FOR_EMAIL: String = "billing_not_verified_for_email"
+}
+
+internal object Constants {
+    const val ACTIVITY_ID_EXTRA_NAME = "activity_id"
 }
